Best school projects/Edge Computing/AI mantrap gate security/fog/web_server/requirements.txt
Flask~=3.1.0
bcrypt~=4.3.0
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
email_validator~=2.2.0
opencv-python~=4.11.0.86
mediapipe~=0.10.18
tensorflow~=2.19.0rc0
librosa~=0.11.0rc1
PyAudio~=0.2.14
SpeechRecognition
paho-mqtt~=2.1.0
flask-cors~=5.0.1
noisereduce~=3.0.3
flask-socketio~=5.5.1
sympy~=1.13.1
